+++ internal/xsql/xquery/rows.go
@@ -11,6 +11,7 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/result"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/types"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/value"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xcontext"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xslices"
 )
@@ -163,6 +164,10 @@ func (r *rows) Next(dst []driver.Value) error {
 
 func (r *rows) Close() error {
 	ctx := context.Background()
+
+	if r.conn != nil && r.conn.ctx != nil {
+		ctx = xcontext.ValueOnly(r.conn.ctx)
+	}
 
 	return r.result.Close(ctx)
 }