+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Fixed several issues in `database/sql` error handling
+* Improved context management for query result streams.
+
 ## v3.113.0
 * Added experimental `config.WithDisableOptimisticUnban` option to disable fast node unban after pessimization
 * Fixed respect start offset from `topicoptions.WithReaderGetPartitionStartOffset` for commit messages